invoice-ai-system/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
python-dotenv==1.0.1
pinecone-client==3.2.2
spacy==3.7.4
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langchain-pinecone==0.1.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pandas==2.2.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
PyMuPDF==1.24.3
